+++ db/migrate/20170308095511_add_notify_projects_channels_to_reminders.rb
@@ -1,0 +1,5 @@
+class AddNotifyProjectsChannelsToReminders < ActiveRecord::Migration[5.0]
+  def change
+    add_column :reminders, :notify_projects_channels, :boolean, null: false, default: false
+  end
+end