+++ spec/use_cases/check_assignments/assign_person_spec.rb
@@ -37,7 +37,7 @@ describe CheckAssignments::AssignPerson do
 
   before do
     allow(project_check).to receive(:decorate) { project_check }
-    allow(project_check).to receive(:slack_channel) { "test" }
+    allow(project_check).to receive(:slack_channels) { "test" }
   end
 
   describe "#call" do